Достигнуто 100% покрытия тестами (по методам)
